Keep a single generic alert card
