--- src/panels/lovelace/cards/security-alerts/context.ts
@@ -1,5 +1,0 @@
-import { createContext } from "@lit/context";
-import type { SecurityAlertItem } from "./helpers";
-
-export const securityAlertsContext =
-  createContext<SecurityAlertItem[]>("security-alerts");
